+++ config/good/webserv.conf
@@ -8,9 +8,9 @@ server {
 	error_page 403 ./www/main/customErrorPages/403.html
 
 	location /toukoum {
+		autoindex on
 		root ./www/
-		index toukoum.html
-		allow_methods POST
+		#index toukoum.html
 	}
 
 	location /alias {
