+++ testers/nginx/nginx-data/sites-available/test.conf
@@ -3,7 +3,7 @@ server {
         # listen   [::]:80 default ipv6only=on; ## listen for ipv6
         root /var/www/test;
 
-        #index index.html index.htm;
+        index index.html index.htm;
 
         error_page 404 505 /404.html;
 
@@ -15,7 +15,8 @@ server {
 				#index nul.html;
 				location /anotherone/ {
 					root /var/www/test/;
-					index index.html;
+					#index index.html;
+					autoindex on;
 				}
 
 				location /alias {
